moving functions to their own files
