+++ src/components/display/fewlines/IdentitySection/IdentitySection.stories.tsx
@@ -1,24 +1,12 @@
 import React from "react";
 
-import { RightChevron } from "../Icons/RightChevron/RightChevron";
-import { IdentityValue } from "../LoginsOverview/LoginsOverview";
 import { IdentitySection } from "./IdentitySection";
-import { Identity, IdentityTypes } from "@lib/@types";
+import { IdentityTypes } from "@lib/@types";
+import { displayIdentity } from "@src/utils/displayIdentity";
 
 export default {
   title: "components/Identity Section",
   component: IdentitySection,
-};
-
-const displayStandardIdentityList = (identity: Identity): JSX.Element => {
-  return (
-    <>
-      <IdentityValue primary={identity.primary} status={identity.status}>
-        {identity.value}
-      </IdentityValue>
-      <RightChevron />
-    </>
-  );
 };
 
 export const EmailIdentitySection = (): JSX.Element => {
@@ -36,7 +24,7 @@ export const EmailIdentitySection = (): JSX.Element => {
             value: "test@test.test",
           },
         ],
-        displayListMethod: displayStandardIdentityList,
+        displayListMethod: displayIdentity,
         disableClick: false,
         noIdentityMessage: "No email added yet.",
         addNewIdentityMessage: "add new email address",
